Set center variable manually for each rule
A context menu Nextcloud Settings window has the same class (Nextcloud)
as the main Nextcloud window, making it indistinguishable from it from
the perspective of hyprland window rule selectors, and centering it
moves it only visually meaning that the click hitbox for the menu
options stays close to the three dots. Having all windows centered makes
it impossible to disable centering only for that one and thus this
change is necessary.